+++ apps/api/src/app/activities/activities.controller.ts
@@ -37,20 +37,24 @@ import {
 } from '@nestjs/common';
 import { REQUEST } from '@nestjs/core';
 import { AuthGuard } from '@nestjs/passport';
-import { Order as OrderModel, Prisma } from '@prisma/client';
+import { Order, Prisma } from '@prisma/client';
 import { parseISO } from 'date-fns';
 import { StatusCodes, getReasonPhrase } from 'http-status-codes';
 
-import { OrderService } from './order.service';
+import { ActivitiesService } from './activities.service';
 
-@Controller('order')
-export class OrderController {
+@Controller([
+  'activities',
+  /** @deprecated */
+  'order'
+])
+export class ActivitiesController {
   public constructor(
+    private readonly activitiesService: ActivitiesService,
     private readonly apiService: ApiService,
     private readonly dataProviderService: DataProviderService,
     private readonly dataGatheringService: DataGatheringService,
     private readonly impersonationService: ImpersonationService,
-    private readonly orderService: OrderService,
     @Inject(REQUEST) private readonly request: RequestWithUser
   ) {}
 
@@ -58,7 +62,7 @@ export class OrderController {
   @HasPermission(permissions.deleteOrder)
   @UseGuards(AuthGuard('jwt'), HasPermissionGuard)
   @UseInterceptors(TransformDataSourceInRequestInterceptor)
-  public async deleteOrders(
+  public async deleteActivities(
     @Query('accounts') filterByAccounts?: string,
     @Query('assetClasses') filterByAssetClasses?: string,
     @Query('dataSource') filterByDataSource?: string,
@@ -73,7 +77,7 @@ export class OrderController {
       filterByTags
     });
 
-    return this.orderService.deleteOrders({
+    return this.activitiesService.deleteActivities({
       filters,
       userId: this.request.user.id
     });
@@ -82,20 +86,20 @@ export class OrderController {
   @Delete(':id')
   @HasPermission(permissions.deleteOrder)
   @UseGuards(AuthGuard('jwt'), HasPermissionGuard)
-  public async deleteOrder(@Param('id') id: string): Promise<OrderModel> {
-    const order = await this.orderService.order({
+  public async deleteActivity(@Param('id') id: string): Promise<Order> {
+    const activity = await this.activitiesService.order({
       id,
       userId: this.request.user.id
     });
 
-    if (!order) {
+    if (!activity) {
       throw new HttpException(
         getReasonPhrase(StatusCodes.FORBIDDEN),
         StatusCodes.FORBIDDEN
       );
     }
 
-    return this.orderService.deleteOrder({
+    return this.activitiesService.deleteActivity({
       id
     });
   }
@@ -105,7 +109,7 @@ export class OrderController {
   @UseInterceptors(RedactValuesInResponseInterceptor)
   @UseInterceptors(TransformDataSourceInRequestInterceptor)
   @UseInterceptors(TransformDataSourceInResponseInterceptor)
-  public async getAllOrders(
+  public async getAllActivities(
     @Headers(HEADER_KEY_IMPERSONATION.toLowerCase()) impersonationId: string,
     @Query('accounts') filterByAccounts?: string,
     @Query('assetClasses') filterByAssetClasses?: string,
@@ -137,7 +141,7 @@ export class OrderController {
       await this.impersonationService.validateImpersonationId(impersonationId);
     const userCurrency = this.request.user.settings.settings.baseCurrency;
 
-    const { activities, count } = await this.orderService.getOrders({
+    const { activities, count } = await this.activitiesService.getActivities({
       endDate,
       filters,
       sortColumn,
@@ -158,7 +162,7 @@ export class OrderController {
   @UseGuards(AuthGuard('jwt'), HasPermissionGuard)
   @UseInterceptors(RedactValuesInResponseInterceptor)
   @UseInterceptors(TransformDataSourceInResponseInterceptor)
-  public async getOrderById(
+  public async getActivityById(
     @Headers(HEADER_KEY_IMPERSONATION.toLowerCase()) impersonationId: string,
     @Param('id') id: string
   ): Promise<ActivityResponse> {
@@ -166,7 +170,7 @@ export class OrderController {
       await this.impersonationService.validateImpersonationId(impersonationId);
     const userCurrency = this.request.user.settings.settings.baseCurrency;
 
-    const { activities } = await this.orderService.getOrders({
+    const { activities } = await this.activitiesService.getActivities({
       userCurrency,
       includeDrafts: true,
       userId: impersonationUserId || this.request.user.id,
@@ -191,7 +195,7 @@ export class OrderController {
   @Post()
   @UseGuards(AuthGuard('jwt'), HasPermissionGuard)
   @UseInterceptors(TransformDataSourceInRequestInterceptor)
-  public async createOrder(@Body() data: CreateOrderDto): Promise<OrderModel> {
+  public async createActivity(@Body() data: CreateOrderDto): Promise<Order> {
     try {
       await this.dataProviderService.validateActivities({
         activitiesDto: [
@@ -227,7 +231,7 @@ export class OrderController {
 
     delete data.dataSource;
 
-    const order = await this.orderService.createOrder({
+    const activity = await this.activitiesService.createActivity({
       ...data,
       date: parseISO(data.date),
       SymbolProfile: {
@@ -252,14 +256,14 @@ export class OrderController {
       userId: this.request.user.id
     });
 
-    if (dataSource && !order.isDraft) {
+    if (dataSource && !activity.isDraft) {
       // Gather symbol data in the background, if data source is set
       // (not MANUAL) and not draft
       this.dataGatheringService.gatherSymbols({
         dataGatheringItems: [
           {
             dataSource,
-            date: order.date,
+            date: activity.date,
             symbol: data.symbol
           }
         ],
@@ -267,19 +271,22 @@ export class OrderController {
       });
     }
 
-    return order;
+    return activity;
   }
 
   @HasPermission(permissions.updateOrder)
   @Put(':id')
   @UseGuards(AuthGuard('jwt'), HasPermissionGuard)
   @UseInterceptors(TransformDataSourceInRequestInterceptor)
-  public async update(@Param('id') id: string, @Body() data: UpdateOrderDto) {
-    const originalOrder = await this.orderService.order({
+  public async updateActivity(
+    @Param('id') id: string,
+    @Body() data: UpdateOrderDto
+  ) {
+    const originalActivity = await this.activitiesService.order({
       id
     });
 
-    if (!originalOrder || originalOrder.userId !== this.request.user.id) {
+    if (!originalActivity || originalActivity.userId !== this.request.user.id) {
       throw new HttpException(
         getReasonPhrase(StatusCodes.FORBIDDEN),
         StatusCodes.FORBIDDEN
@@ -302,7 +309,7 @@ export class OrderController {
 
     delete data.dataSource;
 
-    return this.orderService.updateOrder({
+    return this.activitiesService.updateActivity({
       data: {
         ...data,
         date,