+++ apps/api/src/app/portfolio/portfolio.service.ts
@@ -1,7 +1,7 @@
 import { AccountBalanceService } from '@ghostfolio/api/app/account-balance/account-balance.service';
 import { AccountService } from '@ghostfolio/api/app/account/account.service';
 import { CashDetails } from '@ghostfolio/api/app/account/interfaces/cash-details.interface';
-import { OrderService } from '@ghostfolio/api/app/order/order.service';
+import { ActivitiesService } from '@ghostfolio/api/app/activities/activities.service';
 import { UserService } from '@ghostfolio/api/app/user/user.service';
 import { getFactor } from '@ghostfolio/api/helper/portfolio.helper';
 import { AccountClusterRiskCurrentInvestment } from '@ghostfolio/api/models/rules/account-cluster-risk/current-investment';
@@ -105,13 +105,13 @@ export class PortfolioService {
   public constructor(
     private readonly accountBalanceService: AccountBalanceService,
     private readonly accountService: AccountService,
+    private readonly activitiesService: ActivitiesService,
     private readonly benchmarkService: BenchmarkService,
     private readonly calculatorFactory: PortfolioCalculatorFactory,
     private readonly dataProviderService: DataProviderService,
     private readonly exchangeRateDataService: ExchangeRateDataService,
     private readonly i18nService: I18nService,
     private readonly impersonationService: ImpersonationService,
-    private readonly orderService: OrderService,
     @Inject(REQUEST) private readonly request: RequestWithUser,
     private readonly rulesService: RulesService,
     private readonly symbolProfileService: SymbolProfileService,
@@ -406,7 +406,7 @@ export class PortfolioService {
     const { endDate, startDate } = getIntervalFromDateRange(dateRange);
 
     const { activities } =
-      await this.orderService.getOrdersForPortfolioCalculator({
+      await this.activitiesService.getActivitiesForPortfolioCalculator({
         filters,
         userCurrency,
         userId
@@ -490,7 +490,7 @@ export class PortfolioService {
     );
 
     const { activities } =
-      await this.orderService.getOrdersForPortfolioCalculator({
+      await this.activitiesService.getActivitiesForPortfolioCalculator({
         filters,
         userCurrency,
         userId
@@ -623,6 +623,27 @@ export class PortfolioService {
           ? 0
           : valueInBaseCurrency.div(filteredValueInBaseCurrency).toNumber(),
         assetClass: assetProfile.assetClass,
+        assetProfile: {
+          assetClass: assetProfile.assetClass,
+          assetSubClass: assetProfile.assetSubClass,
+          countries: assetProfile.countries,
+          dataSource: assetProfile.dataSource,
+          holdings: assetProfile.holdings.map(
+            ({ allocationInPercentage, name }) => {
+              return {
+                allocationInPercentage,
+                name,
+                valueInBaseCurrency: valueInBaseCurrency
+                  .mul(allocationInPercentage)
+                  .toNumber()
+              };
+            }
+          ),
+          name: assetProfile.name,
+          sectors: assetProfile.sectors,
+          symbol: assetProfile.symbol,
+          url: assetProfile.url
+        },
         assetSubClass: assetProfile.assetSubClass,
         countries: assetProfile.countries,
         dataSource: assetProfile.dataSource,
@@ -758,7 +779,7 @@ export class PortfolioService {
     const userCurrency = this.getUserCurrency(user);
 
     const { activities } =
-      await this.orderService.getOrdersForPortfolioCalculator({
+      await this.activitiesService.getActivitiesForPortfolioCalculator({
         userCurrency,
         userId
       });
@@ -988,7 +1009,7 @@ export class PortfolioService {
         userId,
         userCurrency
       }),
-      this.orderService.getOrdersForPortfolioCalculator({
+      this.activitiesService.getActivitiesForPortfolioCalculator({
         filters,
         userCurrency,
         userId
@@ -1349,7 +1370,12 @@ export class PortfolioService {
   }) {
     userId = await this.getUserId(impersonationId, userId);
 
-    await this.orderService.assignTags({ dataSource, symbol, tags, userId });
+    await this.activitiesService.assignTags({
+      dataSource,
+      symbol,
+      tags,
+      userId
+    });
   }
 
   private getAggregatedMarkets(holdings: Record<string, PortfolioPosition>): {
@@ -1672,6 +1698,16 @@ export class PortfolioService {
       allocationInPercentage: 0,
       assetClass: AssetClass.LIQUIDITY,
       assetSubClass: AssetSubClass.CASH,
+      assetProfile: {
+        assetClass: AssetClass.LIQUIDITY,
+        assetSubClass: AssetSubClass.CASH,
+        countries: [],
+        dataSource: undefined,
+        holdings: [],
+        name: currency,
+        sectors: [],
+        symbol: currency
+      },
       countries: [],
       dataSource: undefined,
       dateOfFirstActivity: undefined,
@@ -1841,7 +1877,7 @@ export class PortfolioService {
     userId = await this.getUserId(impersonationId, userId);
     const user = await this.userService.user({ id: userId });
 
-    const { activities } = await this.orderService.getOrders({
+    const { activities } = await this.activitiesService.getActivities({
       userCurrency,
       userId,
       withExcludedAccountsAndActivities: true