+++ apps/api/src/services/queues/portfolio-snapshot/portfolio-snapshot.module.ts
@@ -1,5 +1,5 @@
 import { AccountBalanceModule } from '@ghostfolio/api/app/account-balance/account-balance.module';
-import { OrderModule } from '@ghostfolio/api/app/order/order.module';
+import { ActivitiesModule } from '@ghostfolio/api/app/activities/activities.module';
 import { PortfolioCalculatorFactory } from '@ghostfolio/api/app/portfolio/calculator/portfolio-calculator.factory';
 import { CurrentRateService } from '@ghostfolio/api/app/portfolio/current-rate.service';
 import { RedisCacheModule } from '@ghostfolio/api/app/redis-cache/redis-cache.module';
@@ -22,6 +22,7 @@ import { PortfolioSnapshotProcessor } from './portfolio-snapshot.processor';
   exports: [BullModule, PortfolioSnapshotService],
   imports: [
     AccountBalanceModule,
+    ActivitiesModule,
     BullModule.registerQueue({
       name: PORTFOLIO_SNAPSHOT_COMPUTATION_QUEUE,
       settings: {
@@ -36,7 +37,6 @@ import { PortfolioSnapshotProcessor } from './portfolio-snapshot.processor';
     DataProviderModule,
     ExchangeRateDataModule,
     MarketDataModule,
-    OrderModule,
     RedisCacheModule
   ],
   providers: [